fix potential overflow on decimals

+++ contracts/token/ERC20/DetailedERC20.sol
@@ -6,9 +6,9 @@ import "./ERC20.sol";
 contract DetailedERC20 is ERC20 {
   string public name;
   string public symbol;
-  uint8 public decimals;
+  uint256 public decimals;
 
-  function DetailedERC20(string _name, string _symbol, uint8 _decimals) public {
+  function DetailedERC20(string _name, string _symbol, uint256 _decimals) public {
     name = _name;
     symbol = _symbol;
     decimals = _decimals;
